+++ drivers/sensor/liteon/ltr55x/ltr55x.c
@@ -6,85 +6,12 @@
 
 #define DT_DRV_COMPAT liteon_ltr329
 
-#include <zephyr/device.h>
-#include <zephyr/sys/util.h>
-#include <zephyr/drivers/i2c.h>
+#include "ltr55x.h"
+
 #include <zephyr/logging/log.h>
 #include <zephyr/sys/byteorder.h>
-#include <zephyr/drivers/sensor.h>
 
 LOG_MODULE_REGISTER(LTR55X, CONFIG_SENSOR_LOG_LEVEL);
-
-/* Register addresses */
-#define LTR55X_ALS_CONTR      0x80
-#define LTR55X_MEAS_RATE      0x85
-#define LTR55X_PART_ID        0x86
-#define LTR55X_MANUFAC_ID     0x87
-#define LTR55X_ALS_DATA_CH1_0 0x88
-#define LTR55X_ALS_DATA_CH1_1 0x89
-#define LTR55X_ALS_DATA_CH0_0 0x8A
-#define LTR55X_ALS_DATA_CH0_1 0x8B
-#define LTR55X_ALS_STATUS     0x8C
-
-/* Bit masks and shifts for ALS_CONTR register */
-#define LTR55X_ALS_CONTR_MODE_MASK      BIT(0)
-#define LTR55X_ALS_CONTR_MODE_SHIFT     0
-#define LTR55X_ALS_CONTR_SW_RESET_MASK  BIT(1)
-#define LTR55X_ALS_CONTR_SW_RESET_SHIFT 1
-#define LTR55X_ALS_CONTR_GAIN_MASK      GENMASK(4, 2)
-#define LTR55X_ALS_CONTR_GAIN_SHIFT     2
-
-/* Bit masks and shifts for MEAS_RATE register */
-#define LTR55X_MEAS_RATE_REPEAT_MASK    GENMASK(2, 0)
-#define LTR55X_MEAS_RATE_REPEAT_SHIFT   0
-#define LTR55X_MEAS_RATE_INT_TIME_MASK  GENMASK(5, 3)
-#define LTR55X_MEAS_RATE_INT_TIME_SHIFT 3
-
-/* Bit masks and shifts for PART_ID register */
-#define LTR55X_PART_ID_REVISION_MASK  GENMASK(3, 0)
-#define LTR55X_PART_ID_REVISION_SHIFT 0
-#define LTR55X_PART_ID_NUMBER_MASK    GENMASK(7, 4)
-#define LTR55X_PART_ID_NUMBER_SHIFT   4
-
-/* Bit masks and shifts for MANUFAC_ID register */
-#define LTR55X_MANUFAC_ID_IDENTIFICATION_MASK  GENMASK(7, 0)
-#define LTR55X_MANUFAC_ID_IDENTIFICATION_SHIFT 0
-
-/* Bit masks and shifts for ALS_STATUS register */
-#define LTR55X_ALS_STATUS_DATA_MASK        GENMASK(7, 0)
-#define LTR55X_ALS_STATUS_DATA_SHIFT       0
-#define LTR55X_ALS_STATUS_DATA_READY_MASK  BIT(2)
-#define LTR55X_ALS_STATUS_DATA_READY_SHIFT 2
-#define LTR55X_ALS_STATUS_DATA_GAIN_MASK   GENMASK(6, 4)
-#define LTR55X_ALS_STATUS_DATA_GAIN_SHIFT  4
-#define LTR55X_ALS_STATUS_DATA_VALID_MASK  BIT(7)
-#define LTR55X_ALS_STATUS_DATA_VALID_SHIFT 7
-
-/* Expected sensor IDs */
-#define LTR55X_PART_ID_VALUE         0xA0
-#define LTR55X_MANUFACTURER_ID_VALUE 0x05
-
-/* Timing definitions - refer to LTR-329ALS-01 datasheet */
-#define LTR55X_INIT_STARTUP_MS        100
-#define LTR55X_WAKEUP_FROM_STANDBY_MS 10
-
-/* Macros to set and get register fields */
-#define LTR55X_REG_SET(reg, field, value)				\
-	(((value) << reg##_##field##_SHIFT) & reg##_##field##_MASK)
-#define LTR55X_REG_GET(reg, field, value)				\
-	(((value) & reg##_##field##_MASK) >> reg##_##field##_SHIFT)
-
-struct ltr55x_config {
-	const struct i2c_dt_spec bus;
-	uint8_t gain;
-	uint8_t integration_time;
-	uint8_t measurement_rate;
-};
-
-struct ltr55x_data {
-	uint16_t ch0;
-	uint16_t ch1;
-};
 
 static int ltr55x_check_device_id(const struct i2c_dt_spec *bus)
 {
@@ -96,8 +23,8 @@ static int ltr55x_check_device_id(const struct i2c_dt_spec *bus)
 		LOG_ERR("Failed to read PART_ID");
 		return rc;
 	}
-	if (id != LTR55X_PART_ID_VALUE) {
-		LOG_ERR("PART_ID mismatch: expected 0x%02X, got 0x%02X", LTR55X_PART_ID_VALUE, id);
+	if (id != LTR329_PART_ID_VALUE) {
+		LOG_ERR("PART_ID mismatch: expected 0x%02X, got 0x%02X", LTR329_PART_ID_VALUE, id);
 		return -ENODEV;
 	}
